requirements.txt
fastapi
uvicorn
pydantic
qrcode
pillow
psycopg2-binary
python-dotenv
